requirements.txt
ping3
loguru
python-dotenv